AGM: Clean-up of agm service

+++ service/inc/agm_priv.h
@@ -36,7 +36,7 @@
  * Key Vector
  */
 struct agm_key_vector_gsl {
-    uint32_t num_kvs;                 /**< number of key value pairs */
+    size_t num_kvs;                 /**< number of key value pairs */
     struct agm_key_value *kv;       /**< array of key value pairs */
 };
